requirements.txt
uvicorn
fastapi[all]
pymysql
sqlalchemy
bcrypt==4.3.0
passlib[bcrypt]
python-jose[cryptography]